refactor(prepare-binaries): modularize build tool and clean up code formatting
- Extract common utilities, Rari binary, and use-cache addon logic into separate modules (common.rs, rari_binary.rs, use_cache_addon.rs)
- Move shared Target definitions, platform detection, and logging functions to common module
- Simplify main.rs by importing modularized components
- Remove unnecessary braces from single-statement if blocks in native.ts, cache-wrapper.ts for consistency
- Add blank line for readability in deterministic-stringify.ts around Symbol.keyFor logic
- Improve code organization and maintainability while preserving existing functionality
